fix(restore): add Neo4jEnterpriseStandalone support, extract standaloneAsCluster to shared util
